split Summary components to coin specific files
